+++ src/_dependencies/objects/data.py
@@ -1,9 +1,9 @@
 from _dependencies.spec import _Spec
 
 
-def _is_raw(name, dependency):
+def _is_data(name, dependency):
     return True
 
 
-def _build_raw_spec(name, dependency):
+def _build_data_spec(name, dependency):
     return _Spec(lambda: dependency, {}, set(), set())