+++ src/dependencies/__init__.py
@@ -1,8 +1,8 @@
 """Dependency Injection for Humans."""
 from _dependencies.injector import Injector
-from _dependencies.kinds.package import Package
-from _dependencies.kinds.this import this
-from _dependencies.kinds.value import Value as value
+from _dependencies.objects.package import Package
+from _dependencies.objects.this import this
+from _dependencies.objects.value import value
 
 
 __all__ = ["Injector", "Package", "this", "value"]